requirements.txt
requests
speedtest-cli
colorama
tqdm    
netifaces2
psutil